requirements.txt
Appium-Python-Client~=5.2.2
PyYAML~=6.0.2